% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_example}
\alias{data_example}
\title{Example data set with missing values and multilevel struture}
\format{
An object of class \code{data.frame} with 100 rows and 3 columns.
}
\usage{
data_example
}
\description{
This is a generated dataset containing a class variable, a dependent variable y, and
an independent variable X. The data contains missing values in both y and X, assuming a
Missing Completely at Random (MCAR) pattern and a 30% level of missigness.
}
\section{Fields}{

\describe{
\item{\code{y}:}{Object of class \code{"numeric"}, dependent variable}
\item{\code{X}:}{Object of class \code{"numeric"}, independent variable}
\item{\code{class}:}{Object of class \code{"Factor"}, class variable}
}
}

\keyword{datasets}
