% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{normalize}
\alias{normalize}
\title{Normalize dataset}
\usage{
normalize(df_miss)
}
\arguments{
\item{data}{A dataframe with missing values}
}
\value{
A dataframe with normalized values for the numerical variables
}
\description{
This function returns a dataset with normalized values for numerical variables
}
\examples{
data(data_example)
normalize(data_example)
}
