% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{pattern_plot}
\alias{pattern_plot}
\title{Plot pattern of missing values by class}
\usage{
pattern_plot(df, className, varNames)
}
\arguments{
\item{df}{dataframe with missing values}

\item{className}{name of the variable containing classes}

\item{varNames}{vector with the name of the variables to be imputed}
}
\value{
A plot with the patter of missing values by class, by variable
}
\description{
This function returns a dataframe with a multilevel structure. It generates a dataframe using a varying
intercepts/varying slopes linear regression with a single target variable y.
}
\examples{
data(data_example)
pattern_plot(data_example, "class")
}
