% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{plot.missing}
\alias{plot.missing}
\title{Plot number of missing values by class}
\usage{
\method{plot}{missing}(df, className)
}
\arguments{
\item{df}{dataframe with missing values}

\item{className}{name of the variable containing classes}
}
\value{
A barplot with the number of missing values by class, by variable
}
\description{
This function returns a dataframe with a multilevel structure. It generates a dataframe using a varying
intercepts/varying slopes linear regression with a single target variable y.
}
