% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{target_boxplot}
\alias{target_boxplot}
\title{Plot pattern of missing values by class}
\usage{
target_boxplot(df, y, class)
}
\arguments{
\item{df}{dataframe with missing values}

\item{y}{target variable}

\item{class}{name of the variable containing classes}
}
\value{
A boxplot for each class of the target variable
}
\description{
This function returns a dataframe with a multilevel structure. It generates a dataframe using a varying
intercepts/varying slopes linear regression with a single target variable y.
}
\examples{
data(data_example)
target_boxplot(data_example, y, "class")
}
